requirements.txt
Flask
discord.py~=2.6.4
google-generativeai~=0.8.5
google-genai==1.55.0
beautifulsoup4~=4.12.2
ffmpeg
discord-ext-voice-recv
aiohttp~=3.12.14
yt-dlp
yt-search-py
python-dotenv
retry
httpx
mistralai